move ElasticsearchService and ElasticsearchServiceImpl to /es service package
