+++ server/src/main/java/org/tctalent/server/service/db/es/ElasticsearchService.java
@@ -14,7 +14,7 @@
  * along with this program. If not, see https://www.gnu.org/licenses/.
  */
 
-package org.tctalent.server.service.db.impl;
+package org.tctalent.server.service.db.es;
 
 import java.util.Set;
 import org.springframework.lang.NonNull;
