+++ server/src/main/java/org/tctalent/server/service/db/es/ElasticsearchServiceImpl.java
@@ -14,7 +14,7 @@
  * along with this program. If not, see https://www.gnu.org/licenses/.
  */
 
-package org.tctalent.server.service.db.impl;
+package org.tctalent.server.service.db.es;
 
 import io.jsonwebtoken.lang.Collections;
 import java.util.ArrayList;
